pyproject.toml
[project]
name = "research-cli"
version = "0.1.49"
description = "Stateless CLI for Gemini Deep Research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[tool.setuptools]
packages = ["research_cli"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pyinstaller>=6.17.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
research = "research_cli:main"
think = "research_cli:main"
